pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name="mujoco-mjx"
version = "3.2.7"
authors = [
    {name = "Google DeepMind", email = "mujoco@deepmind.com"},
]
description = "MuJoCo XLA (MJX)"
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "Apache License 2.0"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.9"
dependencies = [
    "absl-py",
    "etils[epath]",
    "jax",
    "jaxlib",
    "mujoco>=3.2.7",
    "pytest",
    "scipy",
    "trimesh",
    "warp-lang",
]

[project.scripts]
mjx-testspeed = "mujoco.mjx.testspeed:main"
mjx-viewer = "mujoco.mjx.viewer:main"

[project.urls]
Homepage = "https://github.com/google-deepmind/mujoco/tree/main/mjx"
Documentation = "https://mujoco.readthedocs.io/en/3.2.8"
Repository = "https://github.com/google-deepmind/mujoco/tree/main/mjx"
Changelog = "https://mujoco.readthedocs.io/en/3.2.8/changelog.html"

[tool.isort]
force_single_line = true
force_sort_within_sections = true
lexicographical = true
single_line_exclusions = ["typing"]
order_by_type = false
group_by_package = true
line_length = 120
use_parentheses = true
multi_line_output = 3
skip_glob = ["**/*.ipynb"]

[tool.pyink]
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
extend-exclude = '''(
 .ipynb$
)'''

[tool.ruff]
indent-width = 2